requirements.txt
asgiref==3.3.4
certifi==2020.12.5
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.6
django-bitfield==2.0.1
django-cors-headers==3.1.1
django-dynamic-model-validation==0.1.13
django-filter==2.3.0
django-jsonfield==1.4.0
django-rest-swagger==2.2.0
django-timezone-field==4.0
django-uuidfield==0.5.0
django-xml==2.1.0
djangorestframework==3.10.0
djangorestframework-gis==0.15
djangorestframework-jsonapi==3.1.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djangorestframework-timed-auth-token==1.3.0
drf-yasg==1.17.1
future==0.17.1
geojson==2.5.0
idna==2.10
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
numpy==1.17.3
openapi-codec==1.3.2
packaging==20.9
psycopg2-binary==2.8.6
PyJWT==1.7.1
pyparsing==2.4.7
pyproj==3.0.1
python-dateutil==2.8.1
pytz==2021.1
requests==2.25.1
ruamel.yaml==0.17.5
ruamel.yaml.clib==0.2.2
Shapely==1.7.1
simplejson==3.17.2
six==1.16.0
sqlparse==0.4.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.5
